website/requirements/requirements.txt
Django==1.6.5
gunicorn==19.1.0
South==0.8.4
Werkzeug==0.9.4
django-devserver==0.7.0
httplib2==0.8
oauth2==1.5.211
python-magic==0.4.6
requests==2.2.1
requests-oauthlib==0.4.1
wsgiref==0.1.2
pytz==2014.4
django_compressor==1.4